moved all 'private' methods to _client.py
