+++ tools/pyclient/src/molgenis_emx2_pyclient/_client.py
@@ -1,17 +1,27 @@
+import csv
+import json
 import logging
 import os
+import pathlib
+import time
+from functools import cache
 from typing import Optional
 
+import pandas as pd
 import requests
 from requests import Response
 
 from . import graphql_queries as queries
-from .exceptions import (ServiceUnavailableError, ServerNotFoundError, PyclientException, GraphQLException,
-                         InvalidTokenException,
+from . import utils
+from .constants import HEADING
+from .exceptions import (NoSuchSchemaException, PyclientException,
+                         NoSuchTableException,
+                         GraphQLException, TokenSigninException)
+from .exceptions import (ServiceUnavailableError, ServerNotFoundError, InvalidTokenException,
                          PermissionDeniedException, NonExistentTemplateException,
-                         ReferenceException, NoSuchSchemaException, SigninError, NoContextManagerException,
-                         TokenSigninException)
-from .metadata import Schema
+                         ReferenceException, SigninError, NoContextManagerException)
+from .metadata import Schema, Table
+from .utils import parse_nested_pkeys
 
 logging.getLogger("requests").setLevel(logging.WARNING)
 logging.getLogger("urllib3").setLevel(logging.WARNING)
@@ -74,13 +84,7 @@ class _Client(object):
 
 
     def signin(self, username: str, password: str):
-        """Signs in to the EMX2 server and retrieves session cookie.
-
-        :param username: the username or email address for an account on this server
-        :type username: str
-        :param password: the password corresponding to this username.
-        :type username: str
-        """
+        """Signs in to the EMX2 server."""
         self.username = username
 
         if self.token is not None:
@@ -135,6 +139,8 @@ class _Client(object):
             message = response.json().get('errors')[0].get('message')
             log.error(message)
 
+
+
     def get_schemas(self) -> list[Schema]:
         """Returns the schemas on the database for this user as a list of dictionaries
         containing for each schema the id, name, label and description.
@@ -167,6 +173,62 @@ class _Client(object):
         self.default_schema = name
 
         return name
+
+    @cache
+    def get_schema_metadata(self, name: str = None) -> Schema:
+        """Retrieves a schema's metadata and returns it in a metadata.Schema object."""
+        current_schema = name if name is not None else self.default_schema
+        if current_schema not in self.schema_names:
+            raise NoSuchSchemaException(f"Schema {current_schema!r} not available.")
+
+        query = queries.list_schema_meta()
+        response = self.session.post(
+            url=f"{self.url}/{current_schema}/api/graphql",
+            json={'query': query},
+            headers={'x-molgenis-token': self.token}
+        )
+        self._validate_graphql_response(response)
+
+        response_json = response.json()
+
+        if 'id' not in response_json.get('data').get('_schema'):
+            message = f"Unable to retrieve metadata for schema {current_schema!r}"
+            log.error(message)
+            raise GraphQLException(message)
+
+        metadata = Schema(**response_json.get('data').get('_schema'))
+        return metadata
+
+    def save_schema(self, table: str, name: str = None, file: str = None, data: list | pd.DataFrame = None):
+        """Imports or updates records in a table of a named schema."""
+        current_schema = name
+        if current_schema is None:
+            current_schema = self.default_schema
+
+        if current_schema not in self.schema_names:
+            raise NoSuchSchemaException(f"Schema {current_schema!r} not available.")
+
+        if not self._table_in_schema(table, current_schema):
+            raise NoSuchTableException(f"Table {table!r} not found in schema {current_schema!r}.")
+
+        import_data = self._prep_data_or_file(file_path=file, data=data)
+
+        schema_metadata: Schema = self.get_schema_metadata(current_schema)
+        table_id = schema_metadata.get_table(by='name', value=table).id
+
+        response = self.session.post(
+            url=f"{self.url}/{current_schema}/api/csv/{table_id}",
+            headers={'Content-Type': 'text/csv'},
+            data=import_data.encode('utf-8')
+        )
+
+        try:
+            self._validate_graphql_response(response)
+            log.info("Imported data into %s::%s.", current_schema, table)
+        except PyclientException:
+            errors = '\n'.join([err['message'] for err in response.json().get('errors')])
+            log.error("Failed to import data into %s::%s\n%s", current_schema, table, errors)
+            raise PyclientException(errors)
 
     @property
     def schema_names(self):
@@ -265,3 +327,353 @@ class _Client(object):
             else:
                 message = f"Failed to validate response for {mutation!r}"
                 log.error(message)
+
+    def _upload_csv(self, file_path: pathlib.Path, schema: str) -> str | None:
+        """Uploads the CSV file from the filename to the schema. Returns the success or error message."""
+        file_name = file_path.name
+        if not file_name.startswith('molgenis'):
+            table = file_name.split(file_path.suffix)[0]
+            return self.save_schema(table=table, name=schema, file=str(file_path))
+        api_url = f"{self.url}/{schema}/api/csv"
+        data = self._prep_data_or_file(file_path=str(file_path))
+
+        if self._job:
+            api_url += "?parentJob=" + self._job
+
+        response = self.session.post(
+            url=api_url,
+            data=data,
+            headers={'Content-Type': 'text/csv'}
+        )
+        if response.status_code == 200:
+            msg = response.text
+            log.info(f"{response.text}")
+        else:
+            msg = '\n'.join([err['message'] for err in response.json().get('errors')])
+            log.error(msg)
+            raise PyclientException(msg)
+        return msg
+
+    def _prepare_filter(self, expr: str, _table: str, _schema: str) -> dict | None:
+        """Prepares a GraphQL filter based on the expression passed into `get`."""
+        if expr in [None, ""]:
+            return None
+        statements = expr.split(' and ')
+        _filter = dict()
+        for stmt in statements:
+            if '==' in stmt:
+                _filter.update(**self.__prepare_equals_filter(stmt, _table, _schema))
+            elif '>' in stmt:
+                _filter.update(**self.__prepare_greater_filter(stmt, _table, _schema))
+            elif '<' in stmt:
+                _filter.update(**self.__prepare_smaller_filter(stmt, _table, _schema))
+            elif '!=' in stmt:
+                _filter.update(**self.__prepare_unequal_filter(stmt, _table, _schema))
+            elif 'between' in stmt:
+                _filter.update(**self.__prepare_between_filter(stmt, _table, _schema))
+            else:
+                raise ValueError(f"Cannot process statement {stmt!r}, "
+                                 f"ensure specifying one of the operators '==', '>', '<', '!=', 'between' "
+                                 f"in your statement.")
+        return _filter
+
+    def __prepare_equals_filter(self, stmt: str, _table: str, _schema: str) -> dict:
+        """Prepares the filter part if the statement filters on equality."""
+        _col = stmt.split('==')[0].strip()
+        _val = stmt.split('==')[1].strip()
+
+        col_id = ''.join(_col.split('`'))
+
+        if '.' in col_id:
+            return self.__prepare_nested_filter(col_id, _val, "equals")
+
+        schema = self.get_schema_metadata(_schema)
+        col = schema.get_table(by='name', value=_table).get_column(by='id', value=col_id)
+        match col.get('columnType'):
+            case 'BOOL':
+                val = False
+                if str(_val).lower() == 'true':
+                    val = True
+            case _:
+                try:
+                    val = json.loads(''.join(_val.split('`')).replace("'", '"'))
+                except json.decoder.JSONDecodeError:
+                    val = ''.join(_val.split('`'))
+
+        return {col.id: {'equals': val}}
+
+    def __prepare_greater_filter(self, stmt: str, _table: str, _schema: str) -> dict:
+        """Prepares the filter part if the statement filters on greater than."""
+        exclusive = '=' not in stmt
+        stmt = stmt.replace('=', '')
+
+        _col = stmt.split('>')[0].strip()
+        _val = stmt.split('>')[1].strip()
+
+        col_id = ''.join(_col.split('`'))
+
+        schema = self.get_schema_metadata(_schema)
+        col = schema.get_table(by='name', value=_table).get_column(by='id', value=col_id)
+
+        match col.get('columnType'):
+            case 'INT':
+                val = int(_val) + 1 * exclusive
+            case 'DECIMAL':
+                val = float(_val) + 0.0000001 * exclusive
+            case _:
+                raise NotImplementedError(f"Cannot perform filter '>' on column with type {col.get('columnType')}.")
+
+        return {col.id: {"between": [val, None]}}
+
+    def __prepare_smaller_filter(self, stmt: str, _table: str, _schema: str) -> dict:
+        """Prepares the filter part if the statement filters on greater than."""
+        exclusive = '=' not in stmt
+        stmt = stmt.replace('=', '')
+
+        _col = stmt.split('<')[0].strip()
+        _val = stmt.split('<')[1].strip()
+
+        col_id = ''.join(_col.split('`'))
+
+        schema = self.get_schema_metadata(_schema)
+        col = schema.get_table(by='name', value=_table).get_column(by='id', value=col_id)
+
+        match col.get('columnType'):
+            case 'INT':
+                val = int(_val) - 1 * exclusive
+            case 'DECIMAL':
+                val = float(_val) - 0.0000001 * exclusive
+            case _:
+                raise NotImplementedError(f"Cannot perform filter '<' on column with type {col.get('columnType')}.")
+
+        return {col.id: {"between": [None, val]}}
+
+    def __prepare_unequal_filter(self, stmt: str, _table: str, _schema: str) -> dict:
+        """Prepares the filter part if the statement filters on greater than."""
+        _col = stmt.split('!=')[0].strip()
+        _val = stmt.split('!=')[1].strip()
+
+        col_id = ''.join(_col.split('`'))
+
+        if '.' in col_id:
+            return self.__prepare_nested_filter(col_id, _val, "not_equals")
+
+        schema = self.get_schema_metadata(_schema)
+        col = schema.get_table(by='name', value=_table).get_column(by='id', value=col_id)
+
+        match col.get('columnType'):
+            case _:
+                try:
+                    val = json.loads(''.join(_val.split('`')).replace("'", '"'))
+                except json.decoder.JSONDecodeError:
+                    val = ''.join(_val.split('`'))
+
+        return {col.id: {"not_equals": val}}
+
+    def __prepare_between_filter(self, stmt: str, _table: str, _schema: str) -> dict:
+        """Prepares the filter part if values between a certain range are requested."""
+        stmt.replace('=', '')
+        _col = stmt.split('between')[0].strip()
+        _val = stmt.split('between')[1].strip()
+
+        try:
+            val = json.loads(_val)
+        except json.decoder.JSONDecodeError:
+            msg = ("To filter on values between a and b, supply them as a list, [a, b]. "
+                   "Ensure the values for a and b are numeric.")
+            raise ValueError(msg)
+        col_id = ''.join(_col.split('`'))
+
+        schema = self.get_schema_metadata(_schema)
+        col = schema.get_table(by='name', value=_table).get_column(by='id', value=col_id)
+        if (col_type := col.get('columnType')) not in ['INT', 'DECIMAL']:
+            raise NotImplementedError(f"The filter 'between' is not implemented for columns of type {col_type!r}.")
+
+        return {col.id: {'between': val}}
+
+    def __prepare_nested_filter(self, columns: str, value: str | int | float | list, comparison: str):
+        _filter = {}
+        current = _filter
+        for (i, segment) in enumerate(columns.split('.')[:-1]):
+            current[segment] = {}
+            current = current[segment]
+        last_segment = columns.split('.')[-1]
+        current[last_segment] = {comparison: self.__prepare_value(value)}
+        return _filter
+
+    @staticmethod
+    def __prepare_value(value: str):
+        if value.startswith('[') and value.endswith(']'):
+            return json.loads(value.replace('\'', '"'))
+        return value
+
+    @staticmethod
+    def _prep_data_or_file(file_path: str = None, data: list | pd.DataFrame = None) -> str | None:
+        """Prepares the data from memory or loaded from disk for addition or deletion action.
+
+        :param file_path: path to the file to be prepared
+        :type file_path: str
+        :param data: data to be prepared
+        :type data: list
+
+        :returns: prepared data in dataframe format
+        :rtype: pd.DataFrame
+        """
+
+        if file_path is not None:
+            return utils.read_file(file_path=file_path)
+
+        if data is not None:
+            if isinstance(data, pd.DataFrame):
+                return data.to_csv(index=False, quoting=csv.QUOTE_NONNUMERIC, encoding='UTF-8')
+            else:
+                return pd.DataFrame(data, dtype=str).to_csv(index=False, quoting=csv.QUOTE_NONNUMERIC, encoding='UTF-8')
+
+        message = "No data to import. Specify a file location or a dataset."
+        log.error(message)
+        raise FileNotFoundError(message)
+
+    async def _report_task_progress(self, process_id: int | str):
+        """Reports on the progress of a task and its subtasks."""
+
+        # Report subtask progress
+        p_response = self.session.post(
+            url=self.api_graphql,
+            json={'query': queries.task_status(process_id)}
+        )
+
+        reported_tasks = []
+        task = p_response.json().get('data').get('_tasks')[0]
+        while (status := task.get('status')) != 'COMPLETED':
+            if status == 'ERROR':
+                raise PyclientException(f"Error uploading file: {task.get('description')}")
+            subtasks = task.get('subTasks', [])
+            for st in subtasks:
+                if st['id'] not in reported_tasks and st['status'] == 'RUNNING':
+                    log.info(f"{st['description']}")
+                    reported_tasks.append(st['id'])
+                if st['id'] not in reported_tasks and st['status'] == 'SKIPPED':
+                    log.warning(f"    {st['description']}")
+                    reported_tasks.append(st['id'])
+                for sst in st.get('subTasks', []):
+                    if sst['id'] not in reported_tasks and sst['status'] == 'COMPLETED':
+                        log.info(f"    {sst['description']}")
+                        reported_tasks.append(sst['id'])
+                    if sst['id'] not in reported_tasks and sst['status'] == 'SKIPPED':
+                        log.warning(f"    {sst['description']}")
+                        reported_tasks.append(sst['id'])
+                    for ssst in sst.get('subTasks', []):
+                        if ssst['id'] not in reported_tasks and ssst['status'] == 'COMPLETED':
+                            log.info(f"        {ssst['description']}")
+                            reported_tasks.append(ssst['id'])
+                        if ssst['id'] not in reported_tasks and ssst['status'] == 'SKIPPED':
+                            log.warning(f"        {ssst['description']}")
+                            reported_tasks.append(ssst['id'])
+            try:
+                p_response = self.session.post(
+                    url=self.api_graphql,
+                    json={'query': queries.task_status(process_id)}
+                )
+                task = p_response.json().get('data').get('_tasks')[0]
+            except AttributeError as ae:
+                log.debug(ae)
+                time.sleep(1)
+                p_response = self.session.post(
+                    url=self.api_graphql,
+                    json={'query': queries.task_status(process_id)}
+                )
+                task = p_response.json().get('data').get('_tasks')[0]
+        log.info(f"Completed task: {task.get('description')}")
+
+    @staticmethod
+    def _format_optional_params(**kwargs):
+        """Parses optional keyword arguments to a format suitable for GraphQL queries."""
+        keys = kwargs.keys()
+        args = {key: kwargs[key] for key in keys if (key != 'self') and (key is not None)}
+        if 'name' in args.keys():
+            args['name'] = args.pop('name')
+        if 'include_demo_data' in args.keys():
+            args['includeDemoData'] = args.pop('include_demo_data')
+        if 'parent_job' in args.keys():
+            args['parentJob'] = args.pop('parent_job')
+        return args
+
+    def _table_in_schema(self, table_name: str, schema_name: str) -> bool:
+        """Checks whether the requested table is present in the schema.
+
+        :param table_name: the name of the table
+        :type table_name: str
+        :param schema_name: the name of the schema
+        :type schema_name: str
+        :returns: boolean indicating whether table is present
+        :rtype: bool
+        """
+        schema_data = self.get_schema_metadata(schema_name)
+        if not hasattr(schema_data, 'tables'):
+            return False
+        if table_name in map(str, schema_data.tables):
+            return True
+        return False
+
+    def _parse_get_table_query(self, table_id: str, schema: str, columns: list = None) -> str:
+        """Gathers a table's metadata and parses it to a GraphQL query
+        for querying the table's contents.
+        """
+        schema_metadata: Schema = self.get_schema_metadata(schema)
+        table_metadata: Table = schema_metadata.get_table('id', table_id)
+
+        query = (f"query {table_id}($filter: {table_id}Filter) {{\n"
+                 f"  {table_id}(filter: $filter) {{\n")
+
+        for col in table_metadata.columns:
+            if columns is not None and (col.id not in columns and col.name not in columns):
+                continue
+            if col.get('columnType') in [HEADING]:
+                continue
+            elif col.get('columnType').startswith('ONTOLOGY'):
+                query += f"    {col.get('id')} {{name}}\n"
+            elif col.get('columnType').startswith('REF'):
+                if (ref_schema := col.get('refSchemaName', schema)) == schema:
+                    pkeys = schema_metadata.get_pkeys(col.get('refTableId'))
+                else:
+                    ref_schema_meta = self.get_schema_metadata(ref_schema)
+                    pkeys = ref_schema_meta.get_pkeys(col.get('refTableId'))
+                query += f"    {col.get('id')} {{"
+                query += parse_nested_pkeys(pkeys)
+                query += "}\n"
+            elif col.get('columnType').startswith('FILE'):
+                query += f"    {col.get('id')} {{id}}\n"
+            else:
+                query += f"    {col.get('id')}\n"
+        query += "  }\n"
+        query += "}"
+
+        return query
+
+    def _parse_ontology(self, data: list, table_id: str, schema: str) -> list:
+        """Parses the ontology columns from a GraphQL response."""
+        schema_meta = self.get_schema_metadata(schema)
+        table_meta = schema_meta.get_table('id', table_id)
+        parsed_data = []
+        for row in data:
+            parsed_row = {}
+            for (col, value) in row.items():
+                column_meta = table_meta.get_column('id', col)
+                match column_meta.get('columnType'):
+                    case "ONTOLOGY":
+                        parsed_row[col] = value['name']
+                    case "ONTOLOGY_ARRAY":
+                        parsed_row[col] = [val['name'] for val in value]
+                    case "REF":
+                        _schema = column_meta.get('refSchemaName', schema)
+                        parsed_row[col] = self._parse_ontology([value], column_meta.get('refTableId'), _schema)[0]
+                    case "REF_ARRAY":
+                        _schema = column_meta.get('refSchemaName', schema)
+                        parsed_row[col] = self._parse_ontology(value, column_meta.get('refTableId'), _schema)
+                    case "REFBACK":
+                        _schema = column_meta.get('refSchemaName', schema)
+                        parsed_row[col] = self._parse_ontology(value, column_meta.get('refTableId'), _schema)
+                    case _:
+                        parsed_row[col] = value
+            parsed_data.append(parsed_row)
+        return parsed_data
